Add an 'info' metacommand to produce information
Currently the information is just the shortest path for both players. In the
future there will be recommendation about fences.

This naturally changes the `shortest_distance` function to return an actual
path.

This also simplifies the `shortest_distance` function not to take a type
parameter. Instead of tracking directions at the type level, we just move it to
the value level and compute the abs_diff distance each time. Also fix a small
but serious bug where the heap ordering did not look at the distance already
travelled.