Produce information about defending fences
This is defined as a fence that would prevent the relevant offending fences to
be placed by blocking off an alternative route, or preventing the offedning
fences.

The function `find_obstructing_fence` has been extracted and added a dedup
step: sometimes a winding path can result in multiple identical blocking
fences.

This is still very preliminary because it doesn't attempt to analyze what the
opponent will do after placing this fence.