Replace the `GameStateTree` abstraction with `FenceStateSimulation`
Instead of a tree where the nodes store `GameState` and the edges store
`Fence`s, we simply have a flattened set of `FenceSet` to `[Option<Path>; 2]`.
Here the option would be `None` when the player has no path at all.

This abstraction solves the permutation problem: because it does not
matter the order of placing three fences a,b,c in a simulation, it is
difficult to know whether the tree should store just a,b,c or 3!=6
permutations of those fences. Now we just apply all three fences and then
use the result to compute a path.

Empirically, this is also slightly faster than the previous method.

Also move some checks from `get_potential_defending_fences` into
`evaluate_defending_fences` to reduce duplication.